+++ tests/Unit/Schema/Directives/GuardDirectiveTest.php
@@ -116,4 +116,29 @@ class GuardDirectiveTest extends TestCase
             ],
         ]);
     }
+
+    public function testGuardHappensBeforeOtherDirectivesIfAddedFromType(): void
+    {
+        $this->schema = /** @lang GraphQL */ '
+        type Query @guard {
+            user: User!
+                @can(ability: "adminOnly")
+                @mock
+        }
+
+        type User {
+            name: String
+        }
+        ';
+
+        $this
+            ->graphQL(/** @lang GraphQL */ '
+            {
+                user {
+                    name
+                }
+            }
+            ')
+            ->assertGraphQLErrorCategory(AuthenticationException::CATEGORY);
+    }
 }