+++ tests/Unit/Subscriptions/Broadcasters/PusherBroadcasterTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Unit\Subscriptions\Broadcasters;
 
+use Illuminate\Config\Repository as ConfigRepository;
 use Nuwave\Lighthouse\Subscriptions\BroadcastManager;
 use Nuwave\Lighthouse\Subscriptions\Subscriber;
 use Psr\Log\LoggerInterface;
@@ -17,20 +18,22 @@ class PusherBroadcasterTest extends TestCase
         $logger = $this->createMock(LoggerInterface::class);
 
         // Minimum cases: "create_curl", "trigger POST", "exec_curl response"
-        $logger->expects($this->atLeast(3))->method('log');
+        $logger
+            ->expects($this->atLeast(3))
+            ->method('log');
 
-        // Bind the mocked logger instance instead of the actual logger
-        app()->bind(LoggerInterface::class, function () use ($logger) {
+        $this->app->bind(LoggerInterface::class, function () use ($logger) {
             return $logger;
         });
 
-        // Enable logging for pusher
-        app('config')->set('broadcasting.connections.pusher.log', true);
+        $config = $this->app->make(ConfigRepository::class);
+        $config->set('broadcasting.connections.pusher.log', true);
 
-        $pusherBroadcaster = app(BroadcastManager::class)->driver('pusher');
         $subscriber = $this->createMock(Subscriber::class);
         $subscriber->channel = 'test-123';
 
+        $broadcastManager = $this->app->make(BroadcastManager::class);
+        $pusherBroadcaster = $broadcastManager->driver('pusher');
         $pusherBroadcaster->broadcast($subscriber, 'foo');
     }
 
@@ -38,17 +41,22 @@ class PusherBroadcasterTest extends TestCase
     {
         $logger = $this->createMock(LoggerInterface::class);
 
-        $logger->expects($this->never())->method('log');
+        $logger
+            ->expects($this->never())
+            ->method('log');
 
-        // Bind the mocked logger instance instead of the actual logger
-        app()->bind(LoggerInterface::class, function () use ($logger) {
+        $this->app->bind(LoggerInterface::class, function () use ($logger) {
             return $logger;
         });
 
-        $pusherBroadcaster = app(BroadcastManager::class)->driver('pusher');
+        $config = $this->app->make(ConfigRepository::class);
+        $config->set('broadcasting.connections.pusher.log', false);
+
         $subscriber = $this->createMock(Subscriber::class);
         $subscriber->channel = 'test-123';
 
+        $broadcastManager = $this->app->make(BroadcastManager::class);
+        $pusherBroadcaster = $broadcastManager->driver('pusher');
         $pusherBroadcaster->broadcast($subscriber, 'foo');
     }
 }