+++ CHANGELOG.md
@@ -13,6 +13,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Validate that `@with` and `@withCount` are not used on root fields https://github.com/nuwave/lighthouse/pull/1714
 
+### Changed
+
+- Prepend directives when transferring them from types to fields https://github.com/nuwave/lighthouse/pull/1734
+
 ## 5.2.0
 
 ### Added